+++ submissions/week-2/day-3/Okoye_Nnenna/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "Okoye_Nnenna"
+name = "okoye_nnenna"
 version = "0.1.0"
 edition = "2024"
